[#24] MigratedLoginViewModel 설계 후 View 코드에 바인딩 완료
